+++ src/Oro/Bundle/CMSBundle/Resources/public/js/app/grapesjs/plugins/components/grapesjs-components.js
@@ -8,15 +8,17 @@ import fullscreenCommand from '../../commands/fullscreen';
 import clearCanvasCommand from '../../commands/clear-canvas';
 
 export default GrapesJS.plugins.add('grapesjs-components', function(editor, options) {
-    const superCategoryDefaults = editor.Blocks.Category.prototype.defaults;
-    editor.Blocks.Category.prototype.defaults = () => {
+    const {Blocks, Commands, Panels} = editor;
+
+    const superCategoryDefaults = Blocks.Category.prototype.defaults;
+    Blocks.Category.prototype.defaults = () => {
         return {
             ...superCategoryDefaults(),
             order: 100
         };
     };
 
-    editor.Blocks.Categories = editor.Blocks.Categories.extend({
+    Blocks.Categories = Blocks.Categories.extend({
         comparator: 'order',
 
         add(...args) {
@@ -24,23 +26,6 @@ export default GrapesJS.plugins.add('grapesjs-components', function(editor, opti
             this.sort();
             return res;
         }
-    });
-
-    // Overwrite default addType method
-    // Check and update manually componentTypes array
-    // Check functionality at next version GrapesJS, if need delete wrap function
-    editor.Components.addType = _.wrap(editor.Components.addType, (func, typeName, methods) => {
-        const dom = func.call(editor.Components, typeName, methods);
-
-        const index = _.findIndex(dom.componentTypes, type => type.id === typeName);
-
-        if (index !== -1) {
-            dom.componentTypes[index] = dom.getType(typeName);
-        } else {
-            dom.componentTypes.unshift(dom.getType(typeName));
-        }
-
-        return dom;
     });
 
     editor.getHtml = _.wrap(editor.getHtml, (func, ...args) => unescapeTwigExpression(func.apply(editor, args)));
@@ -53,8 +38,8 @@ export default GrapesJS.plugins.add('grapesjs-components', function(editor, opti
         func.call(editor.editor, opts);
     });
 
-    editor.Commands.add('fullscreen', fullscreenCommand);
-    editor.Commands.add('core:canvas-clear', clearCanvasCommand);
+    Commands.add('fullscreen', fullscreenCommand);
+    Commands.add('core:canvas-clear', clearCanvasCommand);
 
     traitManagerExtends(editor);
 
@@ -65,7 +50,30 @@ export default GrapesJS.plugins.add('grapesjs-components', function(editor, opti
         typeBuildersOptions: _.pick(options, 'excludeContentBlockAlias', 'excludeContentWidgetAlias')
     });
 
-    editor.Panels.removeButton('options', 'preview');
+    Panels.removeButton('options', 'preview');
+    Panels.addButton('options', [
+        {
+            id: 'undo',
+            className: 'fa fa-undo',
+            command(editor) {
+                editor.runCommand('core:undo');
+            }
+        },
+        {
+            id: 'redo',
+            className: 'fa fa-repeat',
+            command(editor) {
+                editor.runCommand('core:redo');
+            }
+        },
+        {
+            id: 'canvas-clear',
+            className: 'fa fa-trash',
+            command(editor) {
+                editor.runCommand('core:canvas-clear');
+            }
+        }
+    ]);
 
     editor.once('destroy', function() {
         editor.componentManager.dispose();