+++ src/Oro/Bundle/CMSBundle/Resources/public/js/app/grapesjs/type-builders/table-tbody-type-builder.js
@@ -19,15 +19,7 @@ const TableTbodyTypeBuilder = BaseTypeBuilder.extend({
     },
 
     isComponent(el) {
-        let result = null;
-
-        if (el.tagName === 'TBODY') {
-            result = {
-                type: this.componentType
-            };
-        }
-
-        return result;
+        return el.nodeType === el.ELEMENT_NODE && el.tagName === 'TBODY';
     }
 });