+++ src/Oro/Bundle/CMSBundle/Resources/public/js/app/grapesjs/type-builders/text-basic-type-builder.js
@@ -1,24 +1,52 @@
-import _ from 'underscore';
 import __ from 'orotranslation/js/translator';
 import BaseTypeBuilder from 'orocms/js/app/grapesjs/type-builders/base-type-builder';
 
 const TextBasicTypeBuilder = BaseTypeBuilder.extend({
+    parentType: 'text',
+
+    button: {
+        label: __('oro.cms.wysiwyg.component.text_basic.label'),
+        category: 'Basic',
+        attributes: {
+            'class': 'fa fa-align-left'
+        }
+    },
+
+    modelMixin: {
+        defaults: {
+            name: __('oro.cms.wysiwyg.component.text_basic.label'),
+            tagName: 'section'
+        },
+
+        init() {
+            const component = this.get('components');
+
+            if (!component.length) {
+                component.add([{
+                    type: 'text',
+                    tagName: 'h1',
+                    components: [{
+                        type: 'textnode',
+                        content: __('oro.cms.wysiwyg.component.text_basic.heading')
+                    }]
+                }, {
+                    type: 'text',
+                    tagName: 'p',
+                    components: [{
+                        type: 'textnode',
+                        content: __('oro.cms.wysiwyg.component.text_basic.paragraph')
+                    }]
+                }]);
+            }
+        }
+    },
+
     constructor: function TextBasicTypeBuilder(options) {
         TextBasicTypeBuilder.__super__.constructor.call(this, options);
     },
 
-    initialize(options) {
-        Object.assign(this, _.pick(options, 'editor', 'componentType'));
-    },
-
-    execute: function() {
-        this.editor.BlockManager.get(this.componentType).set({
-            label: __('oro.cms.wysiwyg.component.text_basic.label'),
-            content: `<section data-gjs-type="text">
-                          <h1>${__('oro.cms.wysiwyg.component.text_basic.heading')}</h1>
-                          <p>${__('oro.cms.wysiwyg.component.text_basic.paragraph')}</p>
-                    </section>`
-        });
+    isComponent() {
+        return false;
     }
 });